Remove unnecessary masternode code and fix logic on version checking
There was code that should have never been included in masternode.cpp that seems to have been causing a segfault on all clients.  There was also a missed condition in the ActiveProtocol check which would have prevented code from executing due to the first condition always proving true.  Thanks to a community member for pointing out something we missed during a rush to get everyone an updated client as quickly as possible.